SoraTgBot/requirements.txt
aiogram>=3.5.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiosqlite>=0.19.0
